+++ InputfieldMystique.module.php
@@ -12,6 +12,8 @@ use Altivebir\Mystique\MystiqueValue;
  * @website			: https://www.altivebir.com
  *
  * @property string $resource
+ * @property bool $useJson
+ * @property string $jsonString
  *
  * @package Altivebir\Mystique
  */
@@ -39,7 +41,7 @@ class InputfieldMystique extends Inputfield {
     {
         return [
             'title' => 'Mystique',
-            'version' => 5,
+            'version' => 6,
             'summary' => __('Provides builder input for ProcessWire CMS/CMF by ALTI VE BIR.'),
             'href' => 'https://www.altivebir.com',
             'author' => 'İskender TOTOĞLU | @ukyo(community), @trk (Github), https://www.altivebir.com',
@@ -63,7 +65,7 @@ class InputfieldMystique extends Inputfield {
 
         $this->wire('classLoader')->addNamespace('Altivebir\Mystique', __DIR__ . '/src');
 
-        $this->resources = Mystique::getResources();
+        $this->resources = Mystique::resources();
 
         $resource = '';
         if(count($this->resources)) {
@@ -72,6 +74,8 @@ class InputfieldMystique extends Inputfield {
 
         // Set default resource
         $this->set('resource', $resource);
+        $this->set('useJson', false);
+        $this->set('jsonString', '');
 	}
 
     /**
@@ -193,6 +197,7 @@ class InputfieldMystique extends Inputfield {
             $mystiqueValue->getPage()->trackChange($this->attr('name'));
         }
 
+        $mystiqueValue->set('__json', $manager->resourceJson);
         $mystiqueValue->set('__resource', $manager->resourceName);
         $mystiqueValue->set('__path', $manager->resourcePath);
 
@@ -205,15 +210,34 @@ class InputfieldMystique extends Inputfield {
      */
     public function ___getConfigInputfields()
     {
-        $resources = Mystique::getResources();
+        $resources = Mystique::resources();
 
         $wrapper = parent::___getConfigInputfields();
+
+        /* @var InputfieldCheckbox $checkbox */
+        $checkbox = $this->wire->modules->get('InputfieldCheckbox');
+        $checkbox->attr('name', 'useJson');
+        $checkbox->set('label', $this->_('Use JSON string'));
+        $checkbox->set('checkboxLabel', $this->_('Use json string instead of a config file.'));
+        $checkbox->attr('checked', $this->useJson ? 'checked' : '');
+
+        $wrapper->append($checkbox);
+
+        /* @var InputfieldTextarea $textarea */
+        $textarea = $this->wire->modules->get('InputfieldTextarea');
+        $textarea->attr('name', 'jsonString');
+        $textarea->set('label', $this->_('JSON string'));
+        $textarea->value = $this->jsonString;
+        $textarea->showIf = "useJson!=''";
+
+        $wrapper->append($textarea);
 
         /* @var InputfieldSelect $select */
         $select = $this->modules->get('InputfieldSelect');
         $select->attr('name', 'resource');
         $select->label = __('Resource');
         $select->required = true;
+        $select->showIf = "useJson=''";
         if(count($resources)) {
             $select->defaultValue = array_keys($resources)[0];
         }