+++ src/MystiqueFormManager.php
@@ -30,6 +30,9 @@ class MystiqueFormManager extends Wire
     /* @var string $name Path of resource */
     public $resourcePath = '';
 
+    /* @var string $resourceJSON Resource data as json string */
+    public $resourceJSON = '';
+
     /* @var string $fields Resource fields */
     public $fields = [];
 
@@ -65,10 +68,15 @@ class MystiqueFormManager extends Wire
         $this->field = $field;
         $this->page = $page;
 
-        $resource =  Mystique::getResource($this->field->resource);
+        if($this->field->useJson && $this->field->jsonString) {
+            $resource = json_decode($field->jsonString, true);
+        } else {
+            $resource =  Mystique::resource($this->field->resource);
+        }
 
-        $this->resourceName = $resource['__name'];
-        $this->resourcePath = $resource['__path'];
+        $this->resourceName = isset($resource['__name']) ? $resource['__name'] : '';
+        $this->resourcePath = isset($resource['__path']) ? $resource['__path'] : '';
+        $this->resourceJSON = json_encode($resource);
         $this->resource = $resource;
 
         $this->setFields($this->resource['fields']);