+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Release notes for Mystique Module
 
+- rename `getResources()` as `resources($json = false)`
+- rename `getResource($name = '')` as `resource($name = '', $json = false)`
+- add option to input field, `use json string` instead of a `config file`
+
+## v.0.0.5
+
 - Fix inside repeater field usage
 
 ## v.0.0.5